Not a number (NaN) handling in every brick action:
* Formula interpretation to NaN will be handled + test.
* Null-Formula tests.
* Stringfunctionhandling of Not-A-Number parameters.

+++ catroidTest/src/org/catrobat/catroid/uitest/formulaeditor/FormulaEditorFragmentTest.java
@@ -322,7 +322,7 @@ public class FormulaEditorFragmentTest extends BaseActivityInstrumentationTestCa
 
 		//Interpretation test
 		Formula formula = (Formula) Reflection.getPrivateField(placeAtBrick, "xPosition");
-		assertEquals("Wrong text in field", newXValue, formula.interpretInteger(sprite));
+		assertEquals("Wrong text in field", newXValue, formula.interpretInteger(sprite).intValue());
 
 		formula = (Formula) Reflection.getPrivateField(placeAtBrick, "yPosition");
